Minor fixes
1. Move fold_cell to smipyping/._common

2. Fix issue that was causing failure in builddoc

3. Update to smiclicmdshelp.rst
